add the shopflow && rating module

+++ database/migrations/2024_10_06_072733_create_comments_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained();
             $table->string('type');
-            $table->integer('data_id');
+            $table->unsignedBigInteger('data_id');
             $table->text('description');
             $table->string('status')->default('not-published');
             $table->timestamps();
