+++ database/migrations/2024_10_06_073001_create_coupons_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('unique_code')->unique();
             $table->string('percent');
-            $table->integer('max_amount');
+            $table->unsignedBigInteger('max_amount');
             $table->integer('max_usage');
             $table->timestamp('started_at');
             $table->timestamp('expire_at');
